+++ nflow-tests/src/main/java/io/nflow/tests/demo/DemoServer.java
@@ -5,6 +5,7 @@ import static io.nflow.engine.workflow.instance.WorkflowInstanceAction.WorkflowA
 import static io.nflow.tests.demo.SpringApplicationContext.applicationContext;
 import static io.nflow.tests.demo.workflow.DemoWorkflow.DEMO_WORKFLOW_TYPE;
 import static io.nflow.tests.demo.workflow.SlowWorkflow.SLOW_WORKFLOW_TYPE;
+import static java.util.Arrays.asList;
 import static java.util.Collections.emptySet;
 import static org.joda.time.DateTime.now;
 
@@ -16,10 +17,13 @@ import org.springframework.context.annotation.Configuration;
 
 import io.nflow.engine.service.WorkflowInstanceInclude;
 import io.nflow.engine.service.WorkflowInstanceService;
+import io.nflow.engine.workflow.definition.BulkWorkflow;
 import io.nflow.engine.workflow.instance.WorkflowInstance;
 import io.nflow.engine.workflow.instance.WorkflowInstanceAction;
+import io.nflow.engine.workflow.instance.WorkflowInstanceFactory;
 import io.nflow.jetty.StartNflow;
 import io.nflow.metrics.NflowMetricsContext;
+import io.nflow.tests.demo.workflow.DemoBulkWorkflow;
 import io.nflow.tests.demo.workflow.DemoWorkflow;
 
 public class DemoServer {
@@ -38,6 +42,7 @@ public class DemoServer {
 
   private static void insertDemoWorkflows() {
     WorkflowInstanceService workflowInstanceService = applicationContext.getBean(WorkflowInstanceService.class);
+    WorkflowInstanceFactory workflowInstanceFactory = applicationContext.getBean(WorkflowInstanceFactory.class);
     WorkflowInstance instance = new WorkflowInstance.Builder().setType(DEMO_WORKFLOW_TYPE)
         .setState(DemoWorkflow.State.begin.name()).build();
     int id = workflowInstanceService.insertWorkflowInstance(instance);
@@ -51,6 +56,14 @@ public class DemoServer {
         .setParentActionId(actionId).setParentWorkflowId(id).build();
     workflowInstanceService.insertWorkflowInstance(child);
     instance = new WorkflowInstance.Builder().setType(SLOW_WORKFLOW_TYPE).setSignal(Optional.of(1)).setNextActivation(null)
+        .build();
+    workflowInstanceService.insertWorkflowInstance(instance);
+    // insert demo bulk workflow with couple of children
+    instance = workflowInstanceFactory.newWorkflowInstanceBuilder() //
+        .setType(DemoBulkWorkflow.DEMO_BULK_WORKFLOW_TYPE) //
+        .setState(BulkWorkflow.State.splitWork.name()) //
+        .putStateVariable(BulkWorkflow.VAR_CONCURRENCY, 2) //
+        .putStateVariable(BulkWorkflow.VAR_CHILD_DATA, asList(1, 2, 3, 4, 5)) //
         .build();
     workflowInstanceService.insertWorkflowInstance(instance);
   }