+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     - h2 1.4.198
 - Fix workflow history cleanup to keep the actions that hold the latest values of state variables
 - nFlow Explorer: Custom content to workflow definition and workflow instance pages. 
+- nFlow Explorer: Executors page to use standard time formatting in tooltips 
 
 ## 5.3.3 (2019-02-04)