+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Allow creating child workflows via REST API (by setting `parentWorkflowId`).
 
 **Details**
+- See `BulkWorkflowTest` and `DemoBulkWorkflow` for examples on how to use bulk workflows
 - Support boxed primitives (Integer, Float etc) with @StateVar
 - nFlow Explorer: Library updates to `lodash` 4.17.11 and `moment` 2.24.0
   Earlier lodash versions had this security vulnerability: https://nvd.nist.gov/vuln/detail/CVE-2018-16487