+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fix workflow history cleanup to keep the actions that hold the latest values of state variables
 - nFlow Explorer: Custom content to workflow definition and workflow instance pages. 
 - nFlow Explorer: Executors page to use standard time formatting in tooltips 
+- nFlow netty: Add support for registering Spring ApplicationListeners
 
 ## 5.3.3 (2019-02-04)